Successfully generate client configs

+++ files/vm-resources/mx/10-openvpn-setup.sh
@@ -15,18 +15,43 @@ fi
 mkdir -p /home/$CUR_USER/programs/openvpn/
 cp openvpn/docker-compose.yml /home/$CUR_USER/programs/openvpn/docker-compose.yaml
 
+hostname=""
 read -p "Please enter your desired hostname (Ex: 10.153.153.15, server.mydomain.com): " hostname
+# echo -n Password: 
+# read -s password
+# echo
+
+if [ "$hostname" == "" ]; then
+    echo "No hostname specified"
+    exit 1
+fi
+
+
+# echo $password
 
 cd /home/$CUR_USER/programs/openvpn
 
 echo "Use same password for everything, leave DN as default (blank)"
+
+# sed -e 's/\s*\([\+0-9a-zA-Z]*\).*/\1/' << EOF | docker-compose run --rm openvpn ovpn_genconfig -u udp://$hostname
+#   $password # password
+#   $password # confirm
+#     # default DN
+#   $password # pass again
+# EOF
+
+# exit 0
 
 docker-compose run --rm openvpn ovpn_genconfig -u udp://$hostname
 docker-compose run --rm openvpn ovpn_initpki
 
 sudo chown -R $(whoami): ./openvpn-data
 
-docker-compose up -d openvpn
+docker-compose up -d
+
+# echo "#!/bin/bash" > up.sh
+# echo "docker-compose up -d openvpn" >> up.sh
+# chmod +x up.sh
 
 echo "#!/bin/bash" > check_logs.sh
 echo "docker-compose logs -f" >> check_logs.sh
@@ -36,20 +61,7 @@ NUM_CLIENTS=2
 
 for i in $(seq 1 $NUM_CLIENTS);
 do
-    echo $i
+    export CLIENTNAME="client$i"
+    docker-compose run --rm openvpn easyrsa build-client-full $CLIENTNAME nopass
+    docker-compose run --rm openvpn ovpn_getclient $CLIENTNAME > $CLIENTNAME.ovpn
 done
-
-export CLIENTNAME="client1"
-docker-compose run --rm openvpn easyrsa build-client-full $CLIENTNAME nopass
-docker-compose run --rm openvpn ovpn_getclient $CLIENTNAME > $CLIENTNAME.ovpn
-
-export CLIENTNAME="client2"
-docker-compose run --rm openvpn easyrsa build-client-full $CLIENTNAME nopass
-docker-compose run --rm openvpn ovpn_getclient $CLIENTNAME > $CLIENTNAME.ovpn
-
-
-# sudo cp openvpn/openvpn /usr/bin/openvpn
-# sudo chmod +x /usr/bin/openvpn
-
-# echo "Installed openvpn to bin"
-# echo "Running 'wireguard' to start the container"
